Split models and utils out of cc.database
This module was getting pretty unwieldy, with the long model
definitions at the top obscuring the actual function definitions. We
move all the code from here into a database subpackage. The only
external change is to a few tests that need to access __DB_VERSION__
and __DEFAULT_DB__ directly (maybe these should be abstracted, but
they also shouldn't be touched normally).